added a bunch of comments and support for a header row

+++ lib/pdf/core.rb
@@ -20,7 +20,7 @@ class Array
   def sum
     s = 0
     each do |v|
-      s += v.to_i
+      s += v.to_f
     end
     s
   end
